Move web server logic to WSC.js
